requirements.txt
alembic==1.10.2
amqp==5.1.1
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
flake8==6.0.0
flake8-bugbear==23.2.13
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-import-order==0.18.2
flake8-variables-names==0.0.5
Flask==2.2.3
Flask-CeleryExt==0.5.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
httplib2==0.21.0
humanize==4.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.1
pandas==1.5.3
pep8-naming==0.13.3
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-mptt==0.2.5
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
